+++ src/main/java/component/SimulatorEvent.java
@@ -4,15 +4,15 @@ package component;
  * Created by dais on 2017-5-17.
  */
 public abstract class SimulatorEvent {
-    private long time;
+    private int time;
 
     public abstract void go();
 
-    public void setTime(long time) {
+    public void setTime(int time) {
         this.time = time;
     }
 
-    public long getTime() {
+    public int getTime() {
         return time;
     }
 }
